+++ extensions/ArrayExtension.js
@@ -1,0 +1,10 @@
+module.exports = class ArrayExtension {
+
+    static shuffle(xs) {
+        for (let i = xs.length - 1; i > 0; i--) {
+            const j = Math.floor(Math.random() * (i + 1));
+            [xs[i], xs[j]] = [xs[j], xs[i]];
+        }
+        return xs;
+    }
+}
